ConvertTabletoDS
YOLO fn requires input data to be in datastore format to use Validation Data, but then documentation claims the user can't shuffle every epoch if input data is a datastore, but it didn't cause an error so I'm not sure if it's shuffling or not.  Added a user warning to that effect.